class 5

+++ my-blog/routes/web.php
@@ -20,12 +20,20 @@ Route::get('/', function () {
 
 use App\Http\Controllers\BlogPostController;
 
-Route::get('blog', [BlogPostController::class, 'index'])->name('blog.index');
-Route::get('blog/{blogPost}', [BlogPostController::class, 'show'])->name('blog.show');
-Route::get('blog-create', [BlogPostController::class, 'create'])->name('blog.create');
-Route::post('blog-create', [BlogPostController::class, 'store']);
-Route::get('blog-edit/{blogPost}', [BlogPostController::class, 'edit'])->name('blog.edit');
-Route::put('blog-edit/{blogPost}', [BlogPostController::class, 'update']);
-Route::delete('blog/{blogPost}', [BlogPostController::class, 'destroy']);
+Route::get('blog', [BlogPostController::class, 'index'])->name('blog.index')->middleware('auth');
+Route::get('blog/{blogPost}', [BlogPostController::class, 'show'])->name('blog.show')->middleware('auth');
+Route::get('blog-create', [BlogPostController::class, 'create'])->name('blog.create')->middleware('auth');
+Route::post('blog-create', [BlogPostController::class, 'store'])->middleware('auth');
+Route::get('blog-edit/{blogPost}', [BlogPostController::class, 'edit'])->name('blog.edit')->middleware('auth');
+Route::put('blog-edit/{blogPost}', [BlogPostController::class, 'update'])->middleware('auth');
+Route::delete('blog/{blogPost}', [BlogPostController::class, 'destroy'])->middleware('auth');
 
 Route::get('query', [BlogPostController::class, 'query']);
+
+use App\Http\Controllers\CustomAuthController;
+Route::get('login', [CustomAuthController::class, 'index'])->name('login');
+Route::post('login', [CustomAuthController::class, 'authentication'])->name('authentication');
+Route::get('registration', [CustomAuthController::class, 'create'])->name('user.registration');
+Route::post('registration', [CustomAuthController::class, 'store'])->name('user.store');
+Route::get('dashboard', [CustomAuthController::class, 'dashboard'])->name('dashboard');
+Route::get('logout', [CustomAuthController::class, 'logout'])->name('logout');
